Resolving circular dependency and refactoring DB Pool in app.js
