+++ package.json
@@ -2,7 +2,7 @@
   "name": "wxr-library",
   "version": "1.6.0",
   "description": "",
-  "main": "app.js",
+  "main": "./server/app.js",
   "scripts": {
     "start": "nodemon ./",
     "debug": "set DEBUG=express:* & nodemon ./",
